refactor: centralize salary component mapping
